# fixes for issue #15 -- ARM assembly alignment bug