--- a/blocks.go
+++ b/blocks.go
@@ -1,4 +1,4 @@
-// +build !amd64 appengine gccgo
+// +build !arm,!amd64 appengine gccgo
 
 package siphash