--- a/blocks_asm.go
+++ b/blocks_asm.go
@@ -1,4 +1,4 @@
-// +build amd64,!appengine,!gccgo
+// +build arm amd64,!appengine,!gccgo
 
 // Written in 2012 by Dmitry Chestnykh.
 //
@@ -12,13 +12,10 @@
 package siphash
 
 //go:noescape
-
-// Hash returns the 64-bit SipHash-2-4 of the given byte slice with two 64-bit
-// parts of 128-bit key: k0 and k1.
-func Hash(k0, k1 uint64, b []byte) uint64
+func blocks(d *digest, p []uint8)
 
 //go:noescape
+func finalize(d *digest) uint64
 
-// Hash128 returns the 128-bit SipHash-2-4 of the given byte slice with two
-// 64-bit parts of 128-bit key: k0 and k1.
-func Hash128(k0, k1 uint64, b []byte) (uint64, uint64)
+//go:noescape
+func once(d *digest)