--- a/hash.go
+++ b/hash.go
@@ -1,4 +1,4 @@
-// +build !amd64 appengine gccgo
+// +build !arm,!amd64 appengine gccgo
 
 // Written in 2012 by Dmitry Chestnykh.
 //