@rakitzis 2018-08-17
@dchest this is a conservative fix for #15. It reintroduces the ARM code streaming hash but omits the HashXXX() family of functions. I would like to look over those more carefully and submit them later, the treatment of the trailing blocks looks okay but I want to err on the side of caution. For my application, anyway, the streaming hash is far more important. I also added an exhaustive check of the unaligned cases in `TestAlign`. This was all verified on a Marvell 3720 (armv7).